requirements.txt
Django==4.2.7
aiogram>=3.20.0
celery==5.3.4
redis==5.0.1
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
django-environ==0.11.2
apscheduler==3.10.4

